Add files via upload
Classification by applying Logistic Regression with the Regularization technique (Lasso (L1), Ridge (L2), and Elastic Net) (Identify the Gender based on the data)

One of the organizations is looking to classify the gender based on the various facial features like nose, forehead, etc. They are having the online business and wanted to implement the same so that if any new customer or individual comes into their website, they will be asked to turn on the camera so that their face measurement can be taken using which the organization wants to create a model. Once a new customer comes in, the machine will be able to predict if the customer is Male or Female. 

I have to apply the regularization technique to solve this problem statement.  

Project Tasks:-
Pre-processing the data:
Missing Value Treatment 
Outlier Treatment if required 
Feature Scaling :
Encoding or any other technique to make categorial to Numerical data
Do extensive exploratory data analysis before creating the model.
Write the detailed conclusions from the analysis. 
Which variables do you think are the most important ones.
Splitting the dataset into train and test sets (if required):
Apply the algorithm which you have learnt, i.e., Logistic Regression and use the concept of regularization.
Applying the K-Fold cross validation technique, you need to find the best value of the C, i.e., 1/Lambda so that the model will not be overfitted or underfitted.
Also, so the model using all the regularization techniques, i.e., Lasso (L1), Ridge (L2), and Elastic net. 
At the end, write the detailed analysis and the behavior you saw from the experiments. 
Finding the best performing model using the accuracy score, confusion matrix, TPR, FPR, TNR, FNR, Precision, Recall, F1-Score, ROC Curve, and loss function. Explain the interpretation of the result.